% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{combPlot}
\alias{combPlot}
\title{Combined plots}
\usage{
combPlot(
  seu,
  features,
  out.dir = "plots",
  plot.name = "combined",
  umap.pt = 2,
  tsne.pt = 2,
  dotplot.cex = 10,
  comb.height = 16,
  comb.width = 20
)
}
\arguments{
\item{seu}{Seurat object}

\item{features}{FindAllMarkers output}

\item{out.dir}{out directory}

\item{plot.name}{file name for plot}

\item{umap.pt}{umap point size}

\item{tsne.pt}{tsne point size}

\item{dotplot.cex}{dotplot axis test size}

\item{comb.height}{plot height}

\item{comb.width}{plot width}
}
\value{
plots in out.dir
}
\description{
Combined plots
}
