% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSeurat.R
\name{CondByIdent}
\alias{CondByIdent}
\title{Compare conditions for each identity class}
\usage{
CondByIdent(
  seu,
  condition = "SHAM.MI",
  xlsx.out.dir = "out",
  xlsx.out.name = "condition-by-ident",
  plot.out.dir = "plots",
  plot.name = "condition-by-ident",
  plot.height = 12,
  plot.width = 16,
  dotplot.cex = 10
)
}
\arguments{
\item{seu}{Seurat Object}

\item{condition}{(metadata column which contains levels of conditions}

\item{xlsx.out.dir}{marker table dir}

\item{xlsx.out.name}{marker table name}

\item{plot.out.dir}{plot dir}

\item{plot.name}{plot name}

\item{plot.height}{plot height}

\item{plot.width}{plot width}

\item{dotplot.cex}{dotplot axis text size}
}
\value{
plots and table of markers
}
\description{
Compare conditions for each identity class
}
