% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ens2gene}
\alias{ens2gene}
\title{Map ensembl IDs to gene names}
\usage{
ens2gene(gcm, edb)
}
\arguments{
\item{gcm}{gene-cell matrix with rownames as genes}

\item{edb}{ensembl database e.g. EnsDb.Mmusculus.v79}
}
\value{
gcm with ensemblID-gene as rownames
}
\description{
Map ensembl IDs to gene names
}
