% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSeurat.R
\name{ExploreClusterResolutions}
\alias{ExploreClusterResolutions}
\title{Visualise clustering parameters}
\usage{
ExploreClusterResolutions(
  seu,
  out.dir = "plots",
  umap.pt = 2,
  tsne.pt = 2,
  dotplot.cex = 10,
  n.markers = 10,
  comb.width = 20,
  comb.height = 16,
  umap.width = 15,
  umap.height = 15,
  tsne.width = 15,
  tsne.height = 15
)
}
\arguments{
\item{seu}{seurat object}

\item{out.dir}{out directory}

\item{umap.pt}{umap point size}

\item{tsne.pt}{tsne point size}

\item{dotplot.cex}{dotplot axis text size}

\item{n.markers}{number of markers}

\item{comb.width}{combined plot width}

\item{comb.height}{combined plot height}

\item{umap.width}{umap plot width}

\item{umap.height}{umap plot height}

\item{tsne.width}{tsne plot width}

\item{tsne.height}{tsne plot height}
}
\value{
list of umaps, tsnes and markers, along with plots in out.dir
}
\description{
Visualise clustering parameters
}
