% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSeurat.R
\name{FindMarkersBaseline}
\alias{FindMarkersBaseline}
\title{Find Markers for all clusters against a baseline cluster}
\usage{
FindMarkersBaseline(
  seu,
  baseline,
  only.pos = TRUE,
  n.markers = 30,
  dotplot.cex = 10
)
}
\arguments{
\item{seu}{Seurat object}

\item{baseline}{cluster to compare all clusters against}

\item{only.pos}{return up-regulated markers against the baseline only}

\item{n.markers}{number of markers to draw in boxplot}

\item{dotplot.cex}{dotplot axis text size}
}
\value{
plots and table of markers
}
\description{
Find Markers for all clusters against a baseline cluster
}
