% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode.vM27.annotation.map}
\alias{gencode.vM27.annotation.map}
\title{Ensembl gene, transcript and common gene name map based on GENCODE vM27}
\format{
A data frame with 142375 rows and 3 columns
Ensembl gene id, ensembl transcript id and common gene name
}
\source{
\url{http://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M27/gencode.vM27.annotation.gtf.gz}
}
\usage{
gencode.vM27.annotation.map
}
\description{
Ensembl gene, transcript and common gene name map based on GENCODE vM27
}
\keyword{datasets}
