% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseRSEM}
\alias{parseRSEM}
\title{Parse RSEM genes.results to a GCM}
\usage{
parseRSEM(path, field = "FPKM")
}
\arguments{
\item{path}{path of directory}

\item{field}{field (FPKM / TPM)}
}
\value{
gcm
}
\description{
Parse RSEM genes.results to a GCM
}
