% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseSTARlog}
\alias{parseSTARlog}
\title{Parse STAR <Sample>.Log.final.out files prefixed by sample name}
\usage{
parseSTARlog(path)
}
\arguments{
\item{path}{path of directory}
}
\value{
mapping statistics
}
\description{
Parse STAR <Sample>.Log.final.out files prefixed by sample name
}
