% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{returnHits}
\alias{returnHits}
\title{Map FindAllMarkers output to a reference}
\usage{
returnHits(features, ref)
}
\arguments{
\item{features}{FindAllMarkers output}

\item{ref}{reference panel}
}
\value{
list of major cell class and tabulated hits
}
\description{
Map FindAllMarkers output to a reference
}
