% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ens2gene}
\alias{ens2gene}
\title{Map ensembl IDs to gene names}
\usage{
ens2gene(gcm, map)
}
\arguments{
\item{gcm}{gene-cell matrix with rownames as genes}

\item{map}{map with ensembl gene id, transcript id and gene name as columns}
}
\value{
gcm with ensemblID-gene as rownames
}
\description{
Map ensembl IDs to gene names
}
